+++ ee/packages/amplication-git-push-webhook-service/src/webhook/webhook.service.ts
@@ -1,24 +1,24 @@
-import { Inject, Injectable, LoggerService } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { EmitterWebhookEventName, Webhooks } from '@octokit/webhooks';
-import { CreateRepositoryPush } from './queue/dto/create-repository-push.dto';
-import { QueueService } from './queue/queue.service';
+import { CreateRepositoryPush } from '../queue/dto/create-repository-push.dto';
+import { QueueService } from '../queue/queue.service';
 import { ConfigService } from '@nestjs/config';
-import { GitOrganizationRepository } from './git-organization/git-organization.repository';
+import { GitOrganizationRepository } from '../git-organization/git-organization.repository';
 import { PushEvent } from '@octokit/webhooks-types';
-import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { EnumProvider } from './git-organization/git-organization.types';
+import { EnumProvider } from '../git-organization/git-organization.types';
+import { AmplicationLogger, AMPLICATION_LOGGER_PROVIDER } from '@amplication/nest-logger-module';
 
 const WEBHOOKS_SECRET_KEY = 'WEBHOOKS_SECRET_KEY';
 
 @Injectable()
-export class AppService {
+export class WebhookService {
   private webhooks: Webhooks;
   constructor(
     private readonly queueService: QueueService,
     configService: ConfigService,
     private readonly gitOrganizationRepository: GitOrganizationRepository,
-    @Inject(WINSTON_MODULE_NEST_PROVIDER)
-    private readonly logger: LoggerService,
+    @Inject(AMPLICATION_LOGGER_PROVIDER)
+    private readonly logger: AmplicationLogger,
   ) {
     this.webhooks = new Webhooks({
       secret: configService.get<string>(WEBHOOKS_SECRET_KEY),
@@ -32,7 +32,7 @@ export class AppService {
     signature: string,
     provider: EnumProvider,
   ) {
-    this.logger.log('start createMessage', { class: AppService.name, id });
+    this.logger.log('start createMessage', { class: WebhookService.name, id });
     switch (eventName.toString().toLowerCase()) {
       case 'push':
         await this.createPushMessage(
@@ -83,7 +83,7 @@ export class AppService {
     if (gitInstallationId) return true;
     this.logger.log(
       `createWebhooksMessage not send, installationId: ${installationId} does not exist`,
-      { class: AppService.name, id },
+      { class: WebhookService.name, id },
     );
     return false;
   }
@@ -104,7 +104,7 @@ export class AppService {
     } catch (error) {
       this.logger.error(
         `failed to createWebhooksMessage: verifyAndReceive, error: ${error}`,
-        { class: AppService.name, id },
+        { class: WebhookService.name, id },
       );
       return false;
     }
@@ -122,7 +122,7 @@ export class AppService {
     if (currentBranch === masterBranch) return true;
     this.logger.log(
       `createWebhooksMessage not send, not master branch, branch name: ${payload.ref}`,
-      { class: AppService.name, id },
+      { class: WebhookService.name, id },
     );
     return false;
   }
